services/assistant/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
orjson>=3.11.1,<3.12
httpx==0.27.0
pydantic-settings==2.5.2
redis[hiredis]==5.0.8
